// src/app/common/model/MaterialAttribute.ts
export class MaterialAttribute{
    Eid:number;
    Id: number;
    MaterialType: string;
    MaterialShortDesc: string;
    MaterialId: string;
    MarketUnitPrice: number;
    Tier: string;
    HoleSection: string;
    HsType: string;
    MGroup: string;
    Od: number;
    Wall: number;
    Weight: number;
    Grade: string;
    Connection: String;
    Vendor: string;
    ManufacturerNum: string;
    SourService: string;
    ManufacturerName: string;
    uoMid: number;
    uom: string;
    vendorSapnumber: string;
    isOdinMaterial:boolean;
}